# server/openapi.yaml
openapi: 3.0.3
info:
  title: Risk Register API
  version: 1.0.0
  description: >-
    REST-API zur Verwaltung eines Risikoregisters.


    Die API unterstützt das Erfassen, Suchen, Aktualisieren und Löschen von
    Risiken.

    Risikowerte liegen zwischen 1 und 25 und ergeben sich fachlich aus
    Eintrittswahrscheinlichkeit × Auswirkung.
  contact:
    name: Risk Register Team
  license:
    name: Private use
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: signed-token
  schemas:
    Risk:
      title: Risk
      description: Ein vollständig gespeichertes Risiko.
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          description: Technische, unveränderliche ID des Risikos.
          example: 01J00000000000000000000001
        reference:
          type: string
          pattern: ^R-\d{3,}$
          description: Menschenlesbare, automatisch vergebene Risikoreferenz.
          example: R-024
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Kurze, eindeutige Bezeichnung des Risikos.
          example: Lieferengpass bei Kernkomponenten
        description:
          type: string
          minLength: 1
          maxLength: 2000
          description: Fachliche Beschreibung des Risikos, seiner Ursache oder möglichen
            Auswirkung.
          example: Risiko von Lieferverzögerungen bei kritischen Kernkomponenten mit
            möglicher Auswirkung auf Produktion und Kundenliefertermine.
        category:
          type: string
          minLength: 1
          maxLength: 200
          description: Fachliche Risikokategorie.
          example: Lieferkette
        owner:
          type: string
          minLength: 1
          maxLength: 200
          description: Für die Behandlung des Risikos verantwortliche Person.
          example: Lena Vogt
        initialScore:
          type: integer
          minimum: 1
          maximum: 25
          description: Risikowert bei der erstmaligen Erfassung
            (Eintrittswahrscheinlichkeit × Auswirkung).
          example: 22
        currentScore:
          type: integer
          minimum: 1
          maximum: 25
          description: Aktuell bewerteter Risikowert.
          example: 18
        status:
          type: string
          enum:
            - Offen
            - In Bearbeitung
            - Überwacht
            - Geschlossen
          description: Aktueller Bearbeitungsstatus des Risikos.
          example: In Bearbeitung
        dueDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Geplantes Fälligkeitsdatum im Format YYYY-MM-DD oder null.
          example: 2026-07-18
        reviewDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Datum der nächsten fachlichen Überprüfung im Format YYYY-MM-DD. Bei
            Fix ist ein Datum erforderlich; bei wiederkehrenden Rhythmen wird es
            serverseitig berechnet.
          example: 2026-07-18
        reviewCycle:
          type: string
          enum:
            - Fix
            - Monatlich
            - Quartalsweise
            - Halbjährlich
            - Jährlich
            - 2-jährlich
          description: Konfigurierbarer Rhythmus für die fachliche Überprüfung. Bei Fix
            ist reviewDate manuell, bei anderen Rhythmen wird reviewDate
            berechnet.
          example: Quartalsweise
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Zeitpunkt der Erfassung im ISO-8601-Format.
          example: 2026-07-01T08:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Zeitpunkt der letzten Änderung im ISO-8601-Format.
          example: 2026-07-07T10:30:00.000Z
        deletedAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          readOnly: true
          description: Zeitpunkt der Löschmarkierung oder null bei aktiven Risiken.
          example: null
      required:
        - id
        - reference
        - title
        - description
        - category
        - owner
        - initialScore
        - currentScore
        - status
        - dueDate
        - reviewDate
        - reviewCycle
        - createdAt
        - updatedAt
        - deletedAt
      example:
        id: 01J00000000000000000000001
        reference: R-024
        title: Lieferengpass bei Kernkomponenten
        description: Risiko von Lieferverzögerungen bei kritischen Kernkomponenten mit
          möglicher Auswirkung auf Produktion und Kundenliefertermine.
        category: Lieferkette
        owner: Lena Vogt
        initialScore: 22
        currentScore: 18
        status: In Bearbeitung
        dueDate: 2026-07-18
        reviewDate: 2026-07-18
        reviewCycle: Quartalsweise
        createdAt: 2026-07-01T08:00:00.000Z
        updatedAt: 2026-07-07T10:30:00.000Z
        deletedAt: null
    CreateRisk:
      title: CreateRisk
      description: Daten zur Erfassung eines neuen Risikos. Nicht gesetzte optionale
        Werte werden aus dem Initialwert abgeleitet.
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Kurze, eindeutige Bezeichnung des Risikos.
          example: Lieferengpass bei Kernkomponenten
        description:
          type: string
          minLength: 1
          maxLength: 2000
          description: Fachliche Beschreibung des Risikos, seiner Ursache oder möglichen
            Auswirkung.
          example: Risiko von Lieferverzögerungen bei kritischen Kernkomponenten mit
            möglicher Auswirkung auf Produktion und Kundenliefertermine.
        category:
          type: string
          minLength: 1
          maxLength: 200
          description: Fachliche Risikokategorie.
          example: Lieferkette
        owner:
          type: string
          minLength: 1
          maxLength: 200
          description: Für die Behandlung des Risikos verantwortliche Person.
          example: Lena Vogt
        initialScore:
          type: integer
          minimum: 1
          maximum: 25
          description: Risikowert bei der erstmaligen Erfassung
            (Eintrittswahrscheinlichkeit × Auswirkung).
          example: 22
        currentScore:
          type: integer
          minimum: 1
          maximum: 25
          description: Aktuell bewerteter Risikowert.
          example: 18
        status:
          type: string
          enum:
            - Offen
            - In Bearbeitung
            - Überwacht
            - Geschlossen
          description: Aktueller Bearbeitungsstatus des Risikos.
          example: In Bearbeitung
        dueDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Geplantes Fälligkeitsdatum im Format YYYY-MM-DD oder null.
          example: 2026-07-18
        reviewDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Datum der nächsten fachlichen Überprüfung im Format YYYY-MM-DD. Bei
            Fix ist ein Datum erforderlich; bei wiederkehrenden Rhythmen wird es
            serverseitig berechnet.
          example: 2026-07-18
        reviewCycle:
          type: string
          enum:
            - Fix
            - Monatlich
            - Quartalsweise
            - Halbjährlich
            - Jährlich
            - 2-jährlich
          description: Konfigurierbarer Rhythmus für die fachliche Überprüfung. Bei Fix
            ist reviewDate manuell, bei anderen Rhythmen wird reviewDate
            berechnet.
          example: Quartalsweise
      required:
        - title
        - description
        - category
        - owner
        - initialScore
      example:
        title: Lieferengpass bei Kernkomponenten
        description: Risiko von Lieferverzögerungen bei kritischen Kernkomponenten mit
          möglicher Auswirkung auf Produktion und Kundenliefertermine.
        category: Lieferkette
        owner: Lena Vogt
        initialScore: 22
        currentScore: 18
        status: In Bearbeitung
        dueDate: 2026-07-18
        reviewDate: 2026-07-18
        reviewCycle: Quartalsweise
    UpdateRisk:
      title: UpdateRisk
      description: Mindestens ein zu änderndes Feld eines vorhandenen Risikos.
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Kurze, eindeutige Bezeichnung des Risikos.
          example: Lieferengpass bei Kernkomponenten
        description:
          type: string
          minLength: 1
          maxLength: 2000
          description: Fachliche Beschreibung des Risikos, seiner Ursache oder möglichen
            Auswirkung.
          example: Risiko von Lieferverzögerungen bei kritischen Kernkomponenten mit
            möglicher Auswirkung auf Produktion und Kundenliefertermine.
        category:
          type: string
          minLength: 1
          maxLength: 200
          description: Fachliche Risikokategorie.
          example: Lieferkette
        owner:
          type: string
          minLength: 1
          maxLength: 200
          description: Für die Behandlung des Risikos verantwortliche Person.
          example: Lena Vogt
        initialScore:
          type: integer
          minimum: 1
          maximum: 25
          description: Risikowert bei der erstmaligen Erfassung
            (Eintrittswahrscheinlichkeit × Auswirkung).
          example: 22
        currentScore:
          type: integer
          minimum: 1
          maximum: 25
          description: Aktuell bewerteter Risikowert.
          example: 18
        status:
          type: string
          enum:
            - Offen
            - In Bearbeitung
            - Überwacht
            - Geschlossen
          description: Aktueller Bearbeitungsstatus des Risikos.
          example: In Bearbeitung
        dueDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Geplantes Fälligkeitsdatum im Format YYYY-MM-DD oder null.
          example: 2026-07-18
        reviewDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Datum der nächsten fachlichen Überprüfung im Format YYYY-MM-DD. Bei
            Fix ist ein Datum erforderlich; bei wiederkehrenden Rhythmen wird es
            serverseitig berechnet.
          example: 2026-07-18
        reviewCycle:
          type: string
          enum:
            - Fix
            - Monatlich
            - Quartalsweise
            - Halbjährlich
            - Jährlich
            - 2-jährlich
          description: Konfigurierbarer Rhythmus für die fachliche Überprüfung. Bei Fix
            ist reviewDate manuell, bei anderen Rhythmen wird reviewDate
            berechnet.
          example: Quartalsweise
      example:
        currentScore: 8
        status: Überwacht
    RiskPage:
      title: RiskPage
      description: Paginierte Ergebnisliste für Risiken.
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Risk"
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 500
          example: 25
        total:
          type: integer
          minimum: 0
          example: 42
        totalPages:
          type: integer
          minimum: 1
          example: 2
      required:
        - items
        - page
        - pageSize
        - total
        - totalPages
    ErrorResponse:
      title: ErrorResponse
      description: Einheitliche Fehlerantwort der Risk-Register-API.
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_NOT_FOUND
            - MEASURE_NOT_FOUND
            - USER_NOT_FOUND
            - AUTH_FAILED
            - FORBIDDEN
            - INTERNAL_ERROR
          example: RISK_NOT_FOUND
        message:
          type: string
          example: Das Risiko mit der angegebenen ID wurde nicht gefunden.
      required:
        - code
        - message
...
paths:
...
  /api/risks:
    get:
      operationId: listRisks
      summary: Risiken abrufen
      tags:
        - Risiken
      description: Liefert Risiken serverseitig gefiltert, sortiert und paginiert.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
          description: Eins-basierte Seitennummer.
        - schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          in: query
          name: pageSize
          required: false
          description: Anzahl der Datensätze pro Seite.
        - schema:
            type: string
            maxLength: 200
          example: Lieferkette
          in: query
          name: search
          required: false
          description: Freitextsuche in Referenz, Bezeichnung, Beschreibung, Kategorie und
            Verantwortlichem.
        - schema:
            type: string
            enum:
              - Offen
              - In Bearbeitung
              - Überwacht
              - Geschlossen
          in: query
          name: status
          required: false
          description: Filtert nach Bearbeitungsstatus.
        - schema:
            type: string
            maxLength: 200
          in: query
          name: category
          required: false
          description: Filtert nach exakt passender Kategorie.
        - schema:
            type: string
            maxLength: 200
          in: query
          name: owner
          required: false
          description: Filtert nach exakt passender verantwortlicher Person.
        - schema:
            type: string
            enum:
              - reference
              - title
              - category
              - owner
              - currentScore
              - status
              - dueDate
              - reviewDate
              - createdAt
              - updatedAt
            default: currentScore
          in: query
          name: sortBy
          required: false
          description: Sortierfeld.
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          in: query
          name: sortDirection
          required: false
          description: Sortierrichtung.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Die gefundene Risikoseite.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskPage"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createRisk
      summary: Ein Risiko erfassen
      tags:
        - Risiken
      description: Erfasst ein neues Risiko. ID, Referenz und Zeitstempel werden
        serverseitig erzeugt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRisk"
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Das erfolgreich erfasste Risiko.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/risks/{id}:
    get:
      operationId: getRisk
      summary: Ein Risiko abrufen
      tags:
        - Risiken
      description: Liefert ein einzelnes Risiko anhand seiner technischen ID.
      parameters:
        - schema:
            type: string
            minLength: 1
          example: 01J00000000000000000000001
          in: path
          name: id
          required: true
          description: Technische ID des Risikos.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Das angeforderte Risiko.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Risiko gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRisk
      summary: Ein Risiko aktualisieren
      tags:
        - Risiken
      description: Aktualisiert die übergebenen Felder eines vorhandenen Risikos.
        Nicht übergebene Felder bleiben unverändert.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRisk"
      parameters:
        - schema:
            type: string
            minLength: 1
          example: 01J00000000000000000000001
          in: path
          name: id
          required: true
          description: Technische ID des Risikos.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Das aktualisierte Risiko.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Risiko gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRisk
      summary: Ein Risiko zur Löschung markieren
      tags:
        - Risiken
      description: Markiert ein Risiko als gelöscht. Es bleibt in SQLite erhalten,
        wird aber in aktiven Listen nicht mehr ausgeliefert.
      parameters:
        - schema:
            type: string
            minLength: 1
          example: 01J00000000000000000000001
          in: path
          name: id
          required: true
          description: Technische ID des Risikos.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Das Risiko wurde erfolgreich zur Löschung markiert.
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Risiko gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
servers:
  - url: http://127.0.0.1:8000
    description: Lokale Entwicklungsumgebung
...